[ARP_SEin_Strukturdaten] Feedback von Oliver fertig eingearbeitet

+++ arp_sein_strukturdaten_pub/process_gemeinden/3_create_json_bodenbedeckungen.sql
@@ -8,15 +8,17 @@ CREATE TABLE
     export.gemeinde_bodenbedeckungen_json AS
     SELECT
         bfs_nr,
-        jsonb_agg(jsonb_build_object(
-            '@type', 'SO_ARP_SEin_Strukturdaten_Publikation_20250407.Strukturdaten.Bodenbedeckung',
-            'Kategorie_Text', kategorie_text,
-            'Flaeche', round(flaeche_agg::NUMERIC, 2)
-        )) AS bodenbedeckungen
+        jsonb_agg(
+            jsonb_build_object(
+                '@type', 'SO_ARP_SEin_Strukturdaten_Publikation_20250407.Strukturdaten.Bodenbedeckung',
+                'Kategorie_Text', kategorie_text,
+                'Flaeche', round(flaeche_agg, 2)
+            )
+        ) AS bodenbedeckungen
     FROM
         export.gemeinde_bodenbedeckung
     WHERE
-        round(flaeche_agg::NUMERIC, 2) > 0
+        round(flaeche_agg, 2) > 0
     GROUP BY
         bfs_nr
 ;
