+++ arp_sein_strukturdaten_pub/process_gemeinden/4_create_json_grundnutzungen.sql
@@ -18,16 +18,20 @@ CREATE TABLE
     )
     SELECT
         bfs_nr,
-        jsonb_agg(jsonb_build_object(
-            '@type', 'SO_ARP_SEin_Strukturdaten_Publikation_20250407.Strukturdaten.Grundnutzung_Kanton',
-            'Kategorie_Text', typ_kt,
-            'Flaeche', round(flaeche::NUMERIC, 2)
-        )) AS grundnutzungen_kanton,
-        jsonb_agg(DISTINCT jsonb_build_object(
-            '@type', 'SO_ARP_SEin_Strukturdaten_Publikation_20250407.Strukturdaten.Grundnutzung_Bund',
-            'Kategorie_Text', typ_bund,
-            'Flaeche', round(flaeche_bund_agg::NUMERIC, 2)
-        )) AS grundnutzungen_bund
+        jsonb_agg(
+            jsonb_build_object(
+                '@type', 'SO_ARP_SEin_Strukturdaten_Publikation_20250407.Strukturdaten.Grundnutzung_Kanton',
+                'Kategorie_Text', typ_kt,
+                'Flaeche', round(flaeche, 2)
+            )
+        ) AS grundnutzungen_kanton,
+        jsonb_agg(
+            DISTINCT jsonb_build_object(
+                '@type', 'SO_ARP_SEin_Strukturdaten_Publikation_20250407.Strukturdaten.Grundnutzung_Bund',
+                'Kategorie_Text', typ_bund,
+                'Flaeche', round(flaeche_bund_agg, 2)
+            )
+        ) AS grundnutzungen_bund
 FROM
     zonentyp_geoms_agg
 GROUP BY
