+++ arp_sein_strukturdaten_pub/process_parzellen/3_create_json_bodenbedeckungen.sql
@@ -10,18 +10,15 @@ CREATE TABLE
         t_ili_tid,
         jsonb_agg(
             jsonb_build_object(
-                '@type',
-                'SO_ARP_SEin_Strukturdaten_Publikation_20250407.Strukturdaten.Bodenbedeckung',
-                'Kategorie_Text',
-                kategorie_text,
-                'Flaeche',
-                round(flaeche_agg::NUMERIC, 2)
+                '@type', 'SO_ARP_SEin_Strukturdaten_Publikation_20250407.Strukturdaten.Bodenbedeckung',
+                'Kategorie_Text', kategorie_text,
+                'Flaeche', round(flaeche_agg, 2)
             )
         ) AS bodenbedeckungen
     FROM
         export.parzellen_bodenbedeckung
     WHERE
-        round(flaeche_agg::NUMERIC, 2) > 0
+        round(flaeche_agg, 2) > 0
     GROUP BY
         t_ili_tid
 ;
