+++ arp_sein_strukturdaten_pub/process_zonenschilde/3_create_json_bodenbedeckungen.sql
@@ -8,15 +8,17 @@ CREATE TABLE
     export.zonenschild_bodenbedeckungen_json AS
     SELECT
         schild_uuid,
-        jsonb_agg(jsonb_build_object(
-            '@type', 'SO_ARP_SEin_Strukturdaten_Publikation_20250407.Strukturdaten.Bodenbedeckung',
-            'Kategorie_Text', kategorie_text,
-            'Flaeche', round(flaeche_agg::NUMERIC, 2)
-        )) AS bodenbedeckungen
+        jsonb_agg(
+            jsonb_build_object(
+                '@type', 'SO_ARP_SEin_Strukturdaten_Publikation_20250407.Strukturdaten.Bodenbedeckung',
+                'Kategorie_Text', kategorie_text,
+                'Flaeche', round(flaeche_agg, 2)
+            )
+        ) AS bodenbedeckungen
     FROM
         export.zonenschild_bodenbedeckung
     WHERE
-        round(flaeche_agg::NUMERIC, 2) > 0
+        round(flaeche_agg, 2) > 0
     GROUP BY
         schild_uuid
 ;
