+++ arp_sein_strukturdaten_pub/process_zonenschilde/8_create_json_statpop.sql
@@ -21,19 +21,21 @@ CREATE TABLE
     )
     SELECT 
         schild_uuid,
-        sum(anzahl) AS popcount,
-        jsonb_agg(jsonb_build_object(
-            '@type', 'SO_ARP_SEin_Strukturdaten_Publikation_20250407.Strukturdaten.Altersklasse_5j',
-            'Kategorie_Id', classagefiveyears,
-            'Kategorie_Text',
-            CASE 
-                WHEN classagefiveyears BETWEEN 0 AND 110 THEN concat(classagefiveyears, '-', classagefiveyears + 4, ' Jahre')
-                WHEN classagefiveyears = 115 THEN '115+ Jahre'
-            END,
-            'Anzahl', anzahl
-        )) altersklassen_5j
+        sum(anzahl)::INTEGER AS popcount,
+        jsonb_agg(
+            jsonb_build_object(
+                '@type', 'SO_ARP_SEin_Strukturdaten_Publikation_20250407.Strukturdaten.Altersklasse_5j',
+                'Kategorie_Id', classagefiveyears,
+                'Kategorie_Text',
+                CASE 
+                    WHEN classagefiveyears BETWEEN 0 AND 110 THEN concat(classagefiveyears, '-', classagefiveyears + 4, ' Jahre')
+                    WHEN classagefiveyears = 115 THEN '115+ Jahre'
+                END,
+                'Anzahl', anzahl
+            )
+        ) altersklassen_5j
     FROM
         statpop_grouped
-GROUP BY
-    schild_uuid
+    GROUP BY
+        schild_uuid
 ;
